feat: Lv3 유저에 비밀번호 필드 추가

+++ src/main/java/com/example/scheduledevelopproject/User/service/UserService.java
@@ -17,9 +17,14 @@ public class UserService {
 
     @Transactional
     public CreateUserResponseDto save(CreateUserRequestDto request) {
+        if (request.getPassword().length() < 8) {
+            throw new IllegalStateException("비밀번호는 8글자 이상이여야 합니다.");
+        }
+
         User user = new User(
                 request.getUserName(),
-                request.getEmail()
+                request.getEmail(),
+                request.getPassword()
         );
 
         User saveUser = userRepository.save(user);
@@ -51,20 +56,27 @@ public class UserService {
                 () -> new IllegalStateException("User whit ID " + userId + "not found.")
         );
 
+        if (!user.getPassword().equals(request.getPassword())) {
+            throw new IllegalStateException("비밀번호가 일치하지 않습니다.");
+        }
+
         user.update(request.getEmail());
 
         return UpdateUserResponseDto.from(user);
     }
 
     @Transactional
-    public void delete(Long userId) {
+    public void delete(Long userId, DeleteUserRequestDto request) {
 
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalStateException("User whit ID " + userId + "not found.")
         );
 
-        userRepository.deleteById(userId);
+        if (!user.getPassword().equals(request.getPassword())) {
+            throw  new IllegalStateException("비밀번호가 일치하지 않습니다.");
+        }
 
+        userRepository.deleteById(userId);
     }
